+++ src/main/java/by/ganevich/io/commands/DeleteClientCommand.java
@@ -7,24 +7,29 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.stereotype.Component;
 
-import java.util.HashMap;
+import java.util.Map;
 
 @Component
 @AllArgsConstructor
 @Getter
 public class DeleteClientCommand implements ICommand {
 
-    private final String commandName = "delete client";
+    private final String commandName = "deleteClient";
 
     private final ClientService clientService;
 
-    //delete client Matvey
     @Override
     public Client execute(CommandDescriptor commandDescriptor) {
 
-        HashMap<Integer, String> parameters = commandDescriptor.getParameters();
+        Map<String, String> parameters = commandDescriptor.getParameters();
 
-        Client client = clientService.findClientByName(parameters.get(0));
+        if (parameters.containsValue("help")){
+            String help = "deleteClient clientName=?";
+            System.out.println(help);
+            return null;
+        }
+
+        Client client = clientService.findClientByName(parameters.get("clientName"));
 
         clientService.removeClient(client);
